# QosService inline metrics reporting