#### Problem
An extra reporting thread is spawned for each QosService instance - which is 1 per banking stage thread. This leads to 6 threads which only do reporting.

- We currently send a `QosMetrics::BlockBatchUpdate { slot: Slot }` every batch to reporting thread via channel
- The separate thread was originally shared and introduced to clean up banking stage and report metrics on an **interval**.
- At later points we:
    - changed these to report per slot which required input from banking threads
    - changed metrics to not be shared between threads

We will use less threads and reduce code complexity by just making banking stage threads report these more directly.

#### Summary of Changes
Just check slot and report QosMetrics inline for banking stage.

Banking bench shows no significant difference in banking performance:
```
with change:
  test bench_banking_stage_multi_accounts             ... bench:  12,391,775 ns/iter (+/- 263,475)
  test bench_banking_stage_multi_accounts_with_voting ... bench:  13,863,025 ns/iter (+/- 844,377)
  test bench_banking_stage_multi_programs             ... bench:  24,131,835 ns/iter (+/- 609,762)
  test bench_banking_stage_multi_programs_with_voting ... bench:  25,511,066 ns/iter (+/- 751,469)

without change:
  test bench_banking_stage_multi_accounts             ... bench:  12,517,154 ns/iter (+/- 144,814)
  test bench_banking_stage_multi_accounts_with_voting ... bench:  13,735,579 ns/iter (+/- 928,524)
  test bench_banking_stage_multi_programs             ... bench:  24,678,824 ns/iter (+/- 363,151)
  test bench_banking_stage_multi_programs_with_voting ... bench:  25,542,058 ns/iter (+/- 1,147,997)
```


Fixes #
<!-- OPTIONAL: Feature Gate Issue: # -->
<!-- Don't forget to add the "feature-gate" label -->