@apfitzge 2023-08-01
> Have you also run this against mnb to ensure there are no latency-related perf issues there too?

Haven't tried on an MNB node, but did run bench-tps w/o any significant effect in numbers (didn't copy them down) both in case w/ no conflicts and with many conflicts. That should be a better test than MNB imo as the transaction load in BS is higher - the only real difference I can think of would be the number of accounts, which has no effect on this change. 

lmk if you'd like me to re-run to grab the numbers of if I should grab a staked mnb node to run on.